### 2019-08-29 #69 @whedon
[PRE REVIEW]: SedEdu: a suite of sediment-related educational modules
**Submitting author:** @amoodie (<a href="http://orcid.org/0000-0002-6745-036X">Andrew Moodie</a>)
**Repository:** <a href="https://github.com/sededu/sededu" target ="_blank">https://github.com/sededu/sededu</a>
**Version:** v1.1.6
