### 2019-08-29 #69 @labarba
👋  @amoodie — After reading your paper, I have some concerns. SedEdu is described as a set of activities, mediated by a software library, for students in the geosciences. Students, however, interact with the content via a GUI, rather than computationally. The focus of JOSE is in using computing (programmatically) to illuminate the learning context/subject. We don't have that here.

Did you want JOSE to review this as a software contribution? (If so, the description as a set of activities doesn't align.) It would not be eligible in the category of learning module, as the learners don't use code directly.
