### 2019-08-29 #69 @amoodie
Hi @labarba. Yes, this is a software submission (I thought I had checked that box on submission?). SedEdu is a software framework that organizes learning activities. That is to say, it is an independent program that could exist without the activities. The framework is lightweight: it is a PyQt GUI that puts submodules into folders. I feel confident that SedEdu is ["open-source software that serves as educational technology"](https://github.com/openjournals/jose/blob/master/README.md#what-do-you-mean-by-open-source-educational-software-tools).

Further, [in a pre-submission inquiry submitted to JOSS](https://github.com/openjournals/joss/issues/537), it was suggested to me that SedEdu could probably fit into JOSS, but was a direct fit for JOSE (of course, I understand that you were not a commenter on that thread). I agree with the commenters that SedEdu fits the spirit of JOSE: I think SedEdu is a really cool project (I am biased of course), and should be highlighted in a journal like JOSE.

I can appreciate your perspective that "the description as a set of activities doesn't align" with this being a software submission, but I think that to describe the software another way would be a disservice to the reader. To elaborate, I am not sure how to tell ["readers...how the content/software has been used, and how they would adopt it,"](https://github.com/openjournals/jose/blob/master/app/views/content/about/_paper_contents.html.erb#L29) ["the "story" of the project,"](https://github.com/openjournals/jose/blob/master/app/views/content/about/_paper_contents.html.erb#L16) or ["the functionality of the software, usage and recent experience of use in teaching and learning situations."](https://github.com/openjournals/jose/blob/master/app/views/content/about/_paper_contents.html.erb#L11) without describing that SedEdu comprises learning activities. I could definitely strip the discussion way down and just describe the core functionality of the software, but this would deprive the reader of understanding the application of the software.

Please advise on how to proceed, cheers.
