### 2019-09-25 #69 @labarba
Hi @amoodie — I'm sorry it's taken so long to get back to you. The beginning of the teaching semester has kept me playing catch-up.

It sounds like you want this submission to be reviewed as a software framework for organizing learning activities. That is, your intended users would be other instructors who want to develop such learning activities (and not the learners themselves). Is that correct?

And in this scenario, the particular set of learning activities that you have developed showcase how the software framework might be used. But you are not submitting the learning activities themselves for review. Correct?

To be clear, as learning activities, my assessment is that they do not quite fit in the JOSE scope (learners do not use coding to learn, they interact with a GUI).

As a software framework to develop learning activities, it would fit the JOSE scope (as educational technology). But in that case, the title of the submission, and the paper itself needs to target the software (not the activities instantiated with the software).

HTH.
