### 2019-09-27 #69 @amoodie
@labarba thanks for getting back to me. You are correct, I want this to be a software submission. I will revise the paper and reply here when I have done so. Thanks.
