### 2019-10-02 f4e8c5d
complete rewrite
