+++ .jose/paper.md
@@ -39,9 +39,9 @@ bibliography: paper.bib
 
 # Summary
 SedEdu is a software framework that organizes computer-based interactive educational activities for sedimentology and stratigraphy courses.
-SedEdu comprises modules which are coupled with activities that guide students, incrementally introducing components of a concept, and testing for understanding and retention throughout the activity.
+Modules are coupled with activities that guide students by incrementally introducing components of a concept, and testing for understanding throughout the activity.
 The modules are computational and quantitative, but the code is abstracted through a graphical user interface (GUI), through which the user interacts with aspects of the computational model.
-This paper focuses on the SedEdu software, and not the underlying modules, an so describes the high level functionality of the SedEdu software package, includes a statement of need for this software package, and provides a brief example of how the SedEdu software has been integrated into a classroom.
+This paper focuses on the SedEdu software, and not the underlying modules, and so describes the high level functionality of the SedEdu software package, includes a statement of need for this software package, and provides a brief example of how the SedEdu software has been integrated into a classroom.
 
 ![SedEdu main screen and category page.](figures/sededu_main_and_category.png)
 
@@ -50,11 +50,10 @@ Clicking on a category button brings the user to that category page, where the a
 Pressing the button "Run module" will launch the module selected, and "Open activity" will open the activity on the users computer (typically a `.pdf` file).
 
 SedEdu is built on PyQt5, and is platform-agnostic, free, fully open source, and available from common software repositories (e.g., `pip`, `conda`).
-SedEdu utilizes `gitsubmodule`s and continuous integration to package and deliver the educational modules, with the goal of making it convenient for users to install, access, and use the software. 
-Modules are developed by individuals, and rely on existing Python libraries, which minimizes the burden of development and maintenance for SedEdu contributors. 
+SedEdu utilizes `gitsubmodule`s and continuous integration to package the educational modules, with the goal of making it convenient for users to install, access, and use the software. 
+Modules are developed by individuals, utilizing existing Python libraries, thereby minimizing the development and maintenance burden for SedEdu contributors. 
 Robust developer instructions included in the repository describe the workflow to develop and integrate a module into SedEdu.
-<!-- SedEdu is modular: modules are developed and maintained by individuals, and are only collated into the framework during building for a new release. -->
-The SedEdu framework is independent of the underlying content material, such that the framework may be adapted (i.e., forked) for use in other fields or communities.
+The SedEdu software is independent of the underlying geologic content material, such that the software may be adapted (i.e., forked) for use in other fields or communities.
 The SedEdu software framework was developed by Andrew J. Moodie, and modules and/or activities have been contributed by Brandee Carlson, Brady Foreman, Andrew Moodie, Kensuke Naito, Jeffrey Kwang, and Jeffrey Nittrouer; there are 7 modules and 7 activities/worksheets in the (current) version v1.1.6, and SedEdu is actively seeking developers of modules and activities.
 
 
@@ -65,12 +64,12 @@ Ideal active learning strategies minimize financial thresholds for entry and max
 Thus, on-line and computer-based learning coupled with in-person lessons, called "blended learning", is a promising avenue to integrate active learning into the classroom [@garrison_blended_2004].
 
 Research and education integration is a core strategy in the U.S.A. National Science Foundation program design [@nsf_merit_2002], so grant writers frequently develop active learning strategies as part of their broader research impacts.
-When developed, these active learning materials are often ineffectively locatable, and so they are underutilized.
-SedEdu organizes these activities into a single place, thereby increasing availability and utilization of externally-funded active learning materials.
+When developed, these active learning materials are often ineffectively disseminated, and so they are underutilized.
+The SedEdu software organizes these activities into a single place, thereby increasing availability and utilization of these externally-funded active learning materials.
 
[…]
