### 2019-10-09 #69 @amoodie
Hi @labarba, I have made revisions to the manuscript, to address your concerns. I believe this software is in line with the JOSE scope as educational infrastructure, and that the paper discusses the contribution in this context. Please let me know whether there is anything else I do, and I look forward to the review and improvement of the software and manuscript. Cheers, AM
