### 2019-11-08 #69 @amoodie
Hi @labarba, I think that @sharondenisse, @lheagy and/or @leouieda will be good reviewers for this paper, because they have listed geoscience as their background. The SedEdu software is written in PyQt; I didn't see anyone who listed this as experience.

Is it possible to have these folks review the submission? Thanks, and please let me know if there is anything I need to do to facilitate this review. Best, AM
