### 2019-11-11 #69 @lheagy
Hi @amoodie, this looks like a really interesting project! Unfortunately I am swamped at the moment, so I can't take on the task of reviewing. You might consider reaching out so some of the folks in the CSDMS group: https://csdms.colorado.edu/wiki/Main_Page, as there seems like a fair bit of overlap in interests. Best of luck with your project!
