### 2019-11-12 #69 @amoodie
@sharondenisse, thanks for your interest in the paper! I am not an editor for JOSE, so I don't have the privilege of initiating the review, or setting a due date. I'm hoping that if some folks are interested in reviewing the paper, the review may be able to begin.

@lheagy, thanks for considering, and for the link!

AM
