### 2019-11-13 #69 @labarba
@amoodie : Did you select the "software" category when you submitted? (if you selected the learning-module category, the review will get formed with the wrong checklist)
