### 2019-12-06 #69 @amoodie
Hi @labarba, is there anything I can do to help facilitate this review?
