### 2019-12-08 #69 @labarba
Thanks for the bump, @amoodie — I have sent two more reviewer enquiries via direct message.
