### 2020-01-26 #69 @amoodie
Hi @labarba, okay, thanks! Here is a short list I scraped together of folks you could try, if you haven't already contacted some of them. Thanks!

@kbarnhart @mcflugen @nicgaspar @saisiddu @nathanlyons @cmshobe @amir-allam @mycarta @jessepisel @mdpiper
