### 2020-02-22 #69 @labarba
Nick @Swanson-Hysell — Would you be interested in contributing a review for The Journal of Open Source Education? https://jose.theoj.org

The submission is titled "SedEdu: a suite of sediment-related educational modules," and it presents a software framework for organizing learning activities, written in PyQt.
[Check out the article proof](https://github.com/openjournals/jose-papers/blob/jose.00069/jose.00069/10.21105.jose.00069.pdf), and the [software repository](https://github.com/sededu/sededu). If you're uninterested or unable to contribute a review, let us know, and then Unsubscribe from this thread to mute any further notifications.

For more information about the journal, see our [documentation](https://openjournals.readthedocs.io/en/jose/) pages.

Thanks! 🙏
