### 2020-02-23 #69 @labarba
👋 @grsharman — Might you be interested in contributing a review for The Journal of Open Source Education? The submission is titled "SedEdu: a suite of sediment-related educational modules," and it presents a software framework for organizing learning activities, written in PyQt.
Check out the [article proof](https://github.com/openjournals/jose-papers/blob/jose.00069/jose.00069/10.21105.jose.00069.pdf), and the [software repository](https://github.com/sededu/sededu).

Read more about the journal in our documentation page: https://openjournals.readthedocs.io/en/jose/
