### 2020-02-24 #69 @Swanson-Hysell
Hi @labarba, I would be happy to serve as a reviewer for this manuscript/project. However, having read through the thread above I saw you make this point:

> It sounds like you want this submission to be reviewed as a software framework for organizing learning activities. That is, your intended users would be other instructors who want to develop such learning activities (and not the learners themselves). Is that correct?

> And in this scenario, the particular set of learning activities that you have developed showcase how the software framework might be used. But you are not submitting the learning activities themselves for review. Correct?

> To be clear, as learning activities, my assessment is that they do not quite fit in the JOSE scope (learners do not use coding to learn, they interact with a GUI).

The way that I would be interested in providing input is in the context of the learning activities. Reviewing it as a "software framework" alone without that disciplinary context effectively means that I would be reviewing the concept of having GUIs that can be accessed from a dialog box for which I wouldn't have much to say. I can see the value in having GUIs wherein the student can vary parameters to run a model under the hood in the context of sedimentology. However, if that type of learning activity doesn't fit with the scope of the journal and isn't what I would be assessing, I am not sure what I would actually be assessing.
