### 2020-02-25 f365b04
add unit test for open and improve error message
