+++ tests/test_utilities.py
@@ -230,11 +230,14 @@ def test_filesList():
 def test_open_file():
     docs_file_list = list(utls.filesList('docs'))
 
-    utls.open_file('docs/'+docs_file_list[0])
+    utls.open_file(os.path.join('docs', docs_file_list[0]))
 
 
 def test_platTypeReading():
-    """Not explicityly in the code as a function"""
+    """
+    Not explicityly in the code as a function.
+    Used to test the Darwin tree path
+    """
     platType = platform.system()
     assert platType in {'Linux', 'Darwin', 'Windows'}
 
@@ -251,8 +254,10 @@ def test_category2path_with_real_names():
 def test_category2path_with_fake_names():
     assert utls.category2path('Andrew Moodie') == 'andrewmoodie'
     assert utls.category2path('SedEdu!IsFun!') == 'sededu!isfun!'
-    # note that this is expected behavior, but 'sededu!isfun!'
-    #    is not a valid file name. It would be best to replace 
-    #    all special characters, or better validate the names?
+    """
+    note that this is expected behavior, but 'sededu!isfun!'
+        is not a valid file name. It would be best to replace 
+        all special characters, or better validate the names?
+    """
 
 
