### 2020-02-26 f8ebd99
jobs add, print info
