+++ .travis.yml
@@ -11,8 +11,11 @@ os:
   - linux
   - osx
 
-osx_image: 
-  - xcode11  
+jobs:
+  include:
+    - language: python
+      python: 3.7
+      osx_image: xcode11  
 
 services:
   - xvfb
@@ -60,6 +63,7 @@ install:
   # - conda install -q numpy scipy matplotlib shapely pillow pytest pyqt
   # - conda install -q -c conda-forge pytest-mpl pytest-qt
   - hash -r
+  - python3 -c "import platform; print('System:', platform.system()); print('Platform:', platform.platform()); print('module v:', platform.__version__)"
 
 before_script:
   # configure a headless display to test plot generation
