+++ .travis.yml
@@ -15,6 +15,7 @@ jobs:
   include:
     - language: python
       python: 3.7
+      os: osx
       osx_image: xcode11  
 
 services:
