+++ .travis.yml
@@ -14,7 +14,6 @@ os:
 jobs:
   include:
     - language: python
-      python: 3.7
       os: osx
       osx_image: xcode11  
 
