### 2020-02-26 3cdcec8
trying to figure out how to build the 10.14 correctly
