### 2020-02-26 971f956
incr ver and expand mat
