+++ .travis.yml
@@ -1,14 +1,14 @@
 env:
   matrix:
-  # - TRAVIS_PYTHON_VERSION="3.5"
-  # - TRAVIS_PYTHON_VERSION="3.6"
+  - TRAVIS_PYTHON_VERSION="3.5"
+  - TRAVIS_PYTHON_VERSION="3.6"
   - TRAVIS_PYTHON_VERSION="3.7"
   global:
   - CONDA_PREFIX=$HOME/miniconda
   - MINICONDA_URL_BASE="https://repo.continuum.io/miniconda/Miniconda3-latest"
 
 os:
-  # - linux
+  - linux
   - osx
 
 jobs:
