### 2020-02-26 2a0006e
remove bad open file test for wind
