### 2020-05-05 #69 @amoodie
Hi @labarba, I was hoping to get an update on the status of this manuscript submission and see if we could identify reviewers. I realize everyone is working double time now with end-of-semester duties and an extra burden from life, but just wanted to provide another gentle nudge! Thanks, AM.
