### 2020-05-28 #69 @jessepisel
Howdy @amoodie and @labarba I have time to review this submission now that the semester has ended. Would I be reviewing it as a software module, or as a learning module as discussed above?
