### 2020-05-29 #69 @amoodie
Thanks @jessepisel, I appreciate your willingness to review! When I submitted, I intended this to be a software submission, but perhaps JOSE's policy has changed since then.

Excited to see what @labarba says, seems like we have two reviewers now?
