### 2020-07-29 #69 @amoodie
Hi @labarba,

I wanted to provide a gentle nudge and see if I could do anything to help facilitate this review. It seems like we may have found two people willing to review, but the window may be closing there, as the summer winds to a close.

Some exciting news is that additional work on SedEdu (including software development and new modules) was recently funded as past of an NSF Postdoc Fellowship that I received. I will begin this work next month, and it would be really great to see v1.x published in JOSE, so that we can use the paper to help advertise to potential collaborators before we begin to work aggressively on the platform.

Thanks very much, hope to not be too much of a nudge here.
