### 2020-10-05 #69 @amoodie
Hi @labarba, is it possible to request an update on the status of this review? With more than 13 months in pre-review, would it be possible to have the submission passed to another editor focused on JOSE reviews at the moment?

Thanks for your time and support in getting our project published in JOSE!
