### 2021-08-28 #69 @Swanson-Hysell
> @Swanson-Hysell — you considered last year contributing a review for the Journal of Open Source Education. Upon our agonizing over the journal scope, I dropped the ball on this one. We now decided to move ahead with the submission as it is intended to be used: learning activities in sediment transport. Might you be able to help us out?

Hi @labarba, yes. I can provide a review associated with the submission. @amoodie has there been progress on the project since your July 29, 2020 message where you indicated that more development was planned? It looks like the last commit was Feb 27, 2020.
