### 2021-08-28 #129 @whedon
[REVIEW]: SedEdu: a suite of sediment-related educational modules
**Submitting author:** @amoodie (<a href="http://orcid.org/0000-0002-6745-036X">Andrew Moodie</a>)
**Repository:** <a href="https://github.com/sededu/sededu" target ="_blank">https://github.com/sededu/sededu</a>
**Version:** v1.1.10a

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode". You can read more about what that means in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss).

- [x] **Version:** v1.1.10a

Conda issue raised here: https://github.com/sededu/sededu/issues/99

The statement of need is very general and can be improved. It cites the benefit of active learning in general without highlighting the level of student that is being targeted or in what context the software is best applied (e.g. in class, in lab, for problem sets).

The conda install is out of date with the current Python version available via Anaconda (https://github.com/sededu/sededu/issues/99). The pip install worked just fine. While I appreciate that conda has made installation relatively straight-forward, I don't know if it rises to the level of simplicity implied here: *Integrating educational infrastructure into the classroom is challenging, in part, due to logistical considerations for installing software; these concerns are alleviated with SedEdu’s simple installation procedure* Given that the software does not expose the students to any code and student interaction is fully through a GUI interface, it is unclear why the developers haven't compiled the software so that it can be downloaded and run as an executable.

Usage is pretty self-explanatory within the GUI and perhaps Fig. 2 can be considered to be example usage.

The example module is probably the best documentation of the software structure. https://github.com/sededu/example-module_sededu It could be valuable to link to it in the paper.

There are tests. I have not engaged with them.

The statement of need is very general and can be improved. It cites the benefit of active learning in general without highlighting the level of student that is being targeted or in what context the software is best applied (e.g. in class, in lab, for problem sets).

- [x] **Version:** v1.1.10a
- [x] #135

* Installation proceeded as outlined. Minor installation issue: sededu/sededu#98 (comment)
* The software makes no performance claims.

* There do not seem to be any software usage examples. This notwithstanding, the software, with its few buttons and sliders, is intuitive to use. Some, but not all, of the modules have worksheets that encourage and guide the the use of relevant sliders. The file [`using_sededu.md`](https://github.com/sededu/sededu/blob/develop/docs/using_sededu.md) is still mostly unpopulated and, thus, does not provide usage examples.
* There is an example of how to write a module to contribute to the software.
* There is an automated test, as part of the Travis CI, that tests the basic functionality of the software but ignores the module tests.
* Some, but not all, modules have test files.
* There does not seem to be instructions detailing manual steps to test the software.
