### 2021-08-28 #129 @labarba
From @Swanson-Hysell [on pre-review](https://github.com/openjournals/jose-reviews/issues/69#issuecomment-907538918):

> @amoodie has there been progress on the project since your July 29, 2020 message where you indicated that more development was planned? It looks like the last commit was Feb 27, 2020.
