### 2021-08-28 #129 @Swanson-Hysell
Hi @labarba, the first review checkbox heads to a 404: https://github.com/openjournals/jose/blob/master/COI.md
where can I find the COI? I don't have a conflict, but don't want to click a box saying that I have read something I haven't.
