### 2021-08-28 #69 @amoodie
@Swanson-Hysell unfortunately no, progress on this has slipped in the chaos of the last year. Eventually we will, though, because SedEdu modules are part of broader impacts on two current NSF Postdoc projects (mine included). Thanks for agreeing to review!
