### 2021-09-24 #129 @hernanrr
Minor installation issue: https://github.com/sededu/rivers2stratigraphy/issues/32#issue-1006850569
