### 2021-09-25 sededu/sededu#98 @hernanrr
Installation issue when calling python3 -m pip install PyQt
Attempting to install from Pypi with Pip, using python3 -m pip install pyqt results in the error shown below.

(vent) % python3 -m pip install pyqt
ERROR: Could not find a version that satisfies the requirement pyqt (from versions: none)
ERROR: No matching distribution found for pyqt
Attempting python3 -m pip install pyqt5 works as expected.

The installation failure seems to be expected behavior, as it is preempted in the installation instructions. If that is the case, why not simply remove it?

Operating system: MacOS Big Sur Version 11.6.
Installation method: Advanced user installations: Install with pip
