### 2021-12-10 sededu/sededu#99 @Swanson-Hysell
conda install requires Python <3.8 which means that installation instructions are out of date
The README tells the user to install Python via Anconda and then conda install the package. However, a user following this installation will get python=3.9 which doesn't appear to be compatible with the sededu conda install.

For example, when I try:
```
conda install -c sededu sededu
```

I get this error:

```
UnsatisfiableError: The following specifications were found
to be incompatible with the existing python installation in your environment:

Specifications:

  - sededu -> python[version='>=3.5,<3.6.0a0|>=3.7,<3.8.0a0|>=3.6,<3.7.0a0']

Your python: python=3.8
```
