### 2021-12-10 #129 @Swanson-Hysell
Regarding installation, I have raised this issue regarding the conda installation:
https://github.com/sededu/sededu/issues/99
