### 2022-02-27 #129 @amoodie
Hi all, yes, I'm working on this today. I'm trying to get the installation issues resolved and taking the opportunity to migrate the CI to github actions too. I'll respond to the reviews here, and the relevant issues, when I'm ready. Thanks reviewers for your time thus far!
