### 2022-02-27 b923233
add actions for conda test and codecov
