+++ requirements.txt
@@ -3,3 +3,5 @@ numpy
 scipy
 matplotlib
 shapely
+pillow
+pyqt5
