+++ .github/workflows/build.yaml
@@ -1,0 +1,46 @@
+# Workflow to automate installation and unit testing of SedEdu on various python versions and operating systems
+
+name: build
+
+on: [push, pull_request]
+
+jobs:
+  test-conda:
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest, macos-latest, windows-latest]
+        python-version: [3.6, 3.7, 3.8, 3.9]
+    env:
+      OS: ${{ matrix.os }}
+      PYTHON: ${{ matrix.python-version }}
+    steps:
+      - uses: conda-incubator/setup-miniconda@v2
+        with:
+          auto-update-conda: true
+          python-version: ${{ matrix.python-version }}
+      - name: Conda info
+        shell: bash -l {0}
+        run: conda info
+      - name: Conda list
+        shell: pwsh
+        run: conda list
+      - name: Install dependencies
+        run: |
+          conda install --file requirements.txt
+          pip install pytest pytest-cov coveralls
+      - name: Install SedEdu
+        run: |
+          pip install -e .
+      - name: Test with pytest
+        run: |
+          pytest --ignore=sededu/modules/ --ignore-glob='*gui.py' --cov=sededu/ --cov-report=xml
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v1
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}
+          file: ./coverage.xml
+          env_vars: OS,PYTHON
+          name: codecov-umbrella
+          fail_ci_if_error: true
