### 2022-02-27 b8b1352
add debug statements
