+++ .github/workflows/build.yaml
@@ -11,7 +11,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: [3.6, 3.7, 3.8, 3.9]
+        python-version: [3.7, 3.8, 3.9]
     env:
       OS: ${{ matrix.os }}
       PYTHON: ${{ matrix.python-version }}
@@ -20,22 +20,24 @@ jobs:
       - uses: conda-incubator/setup-miniconda@v2
         with:
           auto-update-conda: true
+          channels: conda-forge
           python-version: ${{ matrix.python-version }}
+      - name: Install dependencies
+        shell: bash -l {0}
+        run: |
+          conda install --name test --file requirements.txt
+          pip install pytest pytest-mpl pytest-qt pytest-cov coveralls
       - name: Conda info
         shell: bash -l {0}
         run: |
           conda info
           conda list
-      - name: Install dependencies
-        run: |
-          pwd
-          ls -la
-          conda install --file requirements.txt
-          pip install pytest pytest-cov coveralls
       - name: Install SedEdu
+        shell: bash -l {0}
         run: |
           pip install -e .
       - name: Test with pytest
+        shell: bash -l {0}
         run: |
           pytest --ignore=sededu/modules/ --ignore-glob='*gui.py' --cov=sededu/ --cov-report=xml
       - name: Upload coverage to Codecov
