+++ .github/workflows/build.yaml
@@ -16,28 +16,39 @@ jobs:
       OS: ${{ matrix.os }}
       PYTHON: ${{ matrix.python-version }}
     steps:
-      - uses: actions/checkout@v2
-      - uses: conda-incubator/setup-miniconda@v2
+      - name: Checkout SedEdu
+        uses: actions/checkout@v2
+      - name: Setup conda
+        uses: s-weigand/setup-conda@v1
         with:
-          auto-update-conda: true
-          channels: conda-forge
+          update-conda: true
           python-version: ${{ matrix.python-version }}
-      - name: Install dependencies
-        shell: bash -l {0}
+          conda-channels: anaconda, conda-forge
+      - name: Install Python dependencies
         run: |
-          conda install --name test --file requirements.txt
+          conda install --file requirements.txt
           pip install pytest pytest-mpl pytest-qt pytest-cov coveralls
       - name: Conda info
-        shell: bash -l {0}
         run: |
           conda info
           conda list
       - name: Install SedEdu
-        shell: bash -l {0}
         run: |
           pip install -e .
-      - name: Test with pytest
-        shell: bash -l {0}
+      - name: Test with pytest -- osx
+        if: startsWith(matrix.os, 'macOS')
+        run: |
+          pytest --ignore=sededu/modules/ --ignore-glob='*gui.py' --cov=sededu/ --cov-report=xml
+      - name: Test with pytest -- linux
+        if: startsWith(matrix.os, 'ubuntu')
+        run: |
+          export DISPLAY=:99.0
+          sudo apt install libxcb-xinerama0 pyqt5-dev-tools
+          echo "sleeping..."
+          sleep 5
+          xvfb-run `which pytest` --ignore=sededu/modules/ --ignore-glob='*gui.py' --cov=sededu/ --cov-report=xml    
+      - name: Test with pytest -- windows
+        if: startsWith(matrix.os, 'windows')
         run: |
           pytest --ignore=sededu/modules/ --ignore-glob='*gui.py' --cov=sededu/ --cov-report=xml
       - name: Upload coverage to Codecov
