### 2022-02-28 1ac06ac
delete old ci scripts and configs.
