+++ .ci/conda-recipe/build.sh
@@ -1,2 +1,0 @@
-cp -r $RECIPE_DIR/.. $SRC_DIR
-$PYTHON setup.py install --single-version-externally-managed --record record.txt
