+++ .travis.yml
@@ -1,93 +1,0 @@
-env:
-  matrix:
-  - TRAVIS_PYTHON_VERSION="3.5"
-  - TRAVIS_PYTHON_VERSION="3.6"
-  - TRAVIS_PYTHON_VERSION="3.7"
-  global:
-  - CONDA_PREFIX=$HOME/miniconda
-  - MINICONDA_URL_BASE="https://repo.continuum.io/miniconda/Miniconda3-latest"
-
-os:
-  - linux
-  - osx
-
-jobs:
-  include:
-    - TRAVIS_PYTHON_VERSION: "3.5"
-      os: osx
-      osx_image: xcode11  
-
-services:
-  - xvfb
-
-before_install:
-  - echo TRAVIS_BRANCH = $TRAVIS_BRANCH
-  - echo TRAVIS_TAG = $TRAVIS_TAG
-  - echo TRAVIS_PULL_REQUEST = $TRAVIS_PULL_REQUEST
-  - |
-    if [[ $TRAVIS_OS_NAME == "osx" ]]; then
-      brew remove --force $(brew list)
-      brew cleanup -s
-      rm -rf $(brew --cache)
-    fi
-  - |
-    if [[ "$TRAVIS_TAG" == v* ]]; then
-      CIBUILD=1
-      CIRELEASE=1
-    fi
-  - |
-    if [[ "$TRAVIS_BRANCH" == develop ]]; then
-      CIBUILD=1
-    fi
-
-install:
-  - echo "Build on $TRAVIS_OS_NAME for Python $TRAVIS_PYTHON_VERSION"
-  - |
-    if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then
-      OS="MacOSX-x86_64"
-    else
-      OS="Linux-x86_64"
-    fi
-  - curl $MINICONDA_URL_BASE-$OS.sh > $HOME/minconda.sh
-  - bash $HOME/minconda.sh -b -p $CONDA_PREFIX
-  - export PATH="$CONDA_PREFIX/bin:$PATH"
-  - hash -r
-  - conda config --set always_yes yes --set changeps1 no --set anaconda_upload no
-  - conda config --append channels conda-forge
-  - conda update -q conda
-  - conda info -a
-
-  # Create new env and install deps
-  - conda create -q -n test-environment python=$TRAVIS_PYTHON_VERSION numpy scipy matplotlib shapely pillow pyqt pytest>=4.3 pytest-mpl pytest-qt pytest-cov coveralls
-  - source activate test-environment
-  # - conda install -q numpy scipy matplotlib shapely pillow pytest pyqt
-  # - conda install -q -c conda-forge pytest-mpl pytest-qt
-  - hash -r
-  - python3 -c "import platform; print('System:', platform.system()); print('Platform:', platform.platform()); print('module v:', platform.__version__)"
-
-before_script:
-  # configure a headless display to test plot generation
-  - "export DISPLAY=:99.0"
-  # - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then sh -e /etc/init.d/xvfb start ; fi
-  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then ( sudo Xvfb :99 -ac -screen 0 1024x768x8; echo ok )& fi
-  - sleep 5 # give xvfb some time to start
-
-script:
-  # we only test sededu here (NOT submodules), but we do test _opening_ submodules from sededu
-  # - pytest --ignore=sededu/modules/
-  - pytest --ignore=sededu/modules/
-  - if [[ "$TRAVIS_PYTHON_VERSION" == "3.7" ]]; then pytest --ignore=sededu/modules/ --ignore-glob='*gui.py' --cov=sededu/; fi
-
[…]
