+++ .github/workflows/build.yaml
@@ -26,6 +26,7 @@ jobs:
           update-conda: true
           python-version: ${{ matrix.python-version }}
           conda-channels: anaconda, conda-forge
+      
       - name: Install Python dependencies
         run: |
           conda install --file requirements.txt
@@ -34,9 +35,18 @@ jobs:
         run: |
           conda info
           conda list
+      
       - name: Install SedEdu
         run: |
           pip install -e .
+      - name: Set up display -- linux
+        if: startsWith(matrix.os, 'ubuntu')
+        run: |
+          export DISPLAY=:99.0
+          sudo apt install libxcb-xinerama0 pyqt5-dev-tools
+          echo "sleeping..."
+          sleep 5
+      
       - name: Test with pytest -- osx
         if: startsWith(matrix.os, 'macOS')
         run: |
@@ -44,15 +54,12 @@ jobs:
       - name: Test with pytest -- linux
         if: startsWith(matrix.os, 'ubuntu')
         run: |
-          export DISPLAY=:99.0
-          sudo apt install libxcb-xinerama0 pyqt5-dev-tools
-          echo "sleeping..."
-          sleep 5
           xvfb-run `which pytest` --ignore=sededu/modules/ --ignore-glob='*gui.py' --cov=sededu/ --cov-report=xml    
       - name: Test with pytest -- windows
         if: startsWith(matrix.os, 'windows')
         run: |
           pytest --ignore=sededu/modules/ --ignore-glob='*gui.py' --cov=sededu/ --cov-report=xml
+      
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v1
         with:
