### 2022-03-01 ebaede5
add conda-build installation to deploy script
