+++ .github/workflows/deploy.yaml
@@ -91,13 +91,13 @@ jobs:
       - name: Deploy conda package -- osx
         if: startsWith(matrix.os, 'macOS')
         run: |
-          anaconda upload --user sededu --label main .github/conda-build/osx-64/*.tar.bz2
+          anaconda upload -t $ANACONDA_API_TOKEN --user sededu --label main .github/conda-build/osx-64/*.tar.bz2
       - name: Deploy conda package -- linux
         if: startsWith(matrix.os, 'ubuntu')
         run: |
-          anaconda upload --user sededu --label main .github/conda-build/linux-64/*.tar.bz2
+          anaconda upload -t $ANACONDA_API_TOKEN --user sededu --label main .github/conda-build/linux-64/*.tar.bz2
       - name: Deploy conda package -- windows
         if: startsWith(matrix.os, 'windows')
         run: |
-          anaconda upload --user sededu --label main .github/conda-build/win-64/*.tar.bz2
+          anaconda upload -t $ANACONDA_API_TOKEN --user sededu --label main .github/conda-build/win-64/*.tar.bz2
      
