### 2022-03-01 0bd8030
update readme with build status, increment version, update deploy script.
