### 2022-03-02 6a851ab
update readme documentation and version num
