+++ sededu/_version.py
@@ -1,1 +1,1 @@
-__version__ = "1.1.8c"
+__version__ = "1.1.9"
