### 2022-03-21 bfcbe69
update various installation and setup reqts and instructions.
