### 2022-03-21 0eda65e
update conda installation to use environment.yml file
