+++ .github/workflows/build.yaml
@@ -29,7 +29,7 @@ jobs:
       
       - name: Install Python dependencies
         run: |
-          conda install --file requirements.txt
+          conda env update --file environment.yml
           pip install pytest pytest-mpl pytest-qt pytest-cov coveralls
       - name: Conda info
         run: |
