### 2022-03-21 9175647
update conda installation to use environment.yml file
