+++ .github/workflows/build.yaml
@@ -29,7 +29,7 @@ jobs:
       
       - name: Install Python dependencies
         run: |
-          conda env update --file environment.yml
+          conda env update -n base --file environment.yml
           pip install pytest pytest-mpl pytest-qt pytest-cov coveralls
       - name: Conda info
         run: |
