+++ docs/using_sededu.md
@@ -1,12 +1,15 @@
 # Using SedEdu
 
-* This page will contain a walkthrough of the layout and functionality of SedEdu.
-* It is intended for end-users and should be simple and tech-jargon free.
+![SedEdu main screen and category page.](https://github.com/sededu/sededu/raw/develop/.jose/figures/sededu_main_and_category.png)
+
+* The SedEdu framework organizes modules into five categories (Figure 1, left): "Rivers and deltas", "Landscapes", "Deserts", "Coasts", and "Stratigraphy". 
+* Clicking on a category button brings the user to that category page, where the available modules (and associated metadata and activities) are listed (Figure 1, right).
+* Pressing the button "Run module" will launch the module selected, and "Open activity" will open the activity on the users computer (typically a .pdf file).
 
 
 ## Some ideas on ways to use it
 
-* home-schooling
 * on the board, used by instructor
 * on class computers with an activity
-* substitute teacher activity 
+* home-schooling
+* problem sets for homework
