### 2022-03-22 sededu/sededu#99 @amoodie
Thank you for pointing this out. Installation instructions are updated with v1.1.10.

I've tested installation:
 * continuous integratation on GitHub Actions, building and testing on linux, windows, and macos with Python 3.7, 3.8, 3.9
 * Manually tested installation with my own linux and windows machines, following installation instructions in the README
