### 2022-03-22 sededu/sededu#98 @amoodie
Thank you for pointing this out. Installation instructions are updated with v1.1.10.

Installation for `pyqt` is tricky, because the package names change with major version (e.g., `pyqt4`, vs `pyqt5`) and differ between platforms. For example, on `conda`, `pyqt` will install the latest `pyqt5` version, but on `pip` you need to name the version (as you identified).

I've revised the installation instructions to reflect this discrepancy if `pip` installing, and added an `environment.yml` file to make the `pip` `requirements.txt` separate from the `conda` requirements list.
