### 2022-03-22 #129 @amoodie
# summary

Thank you to the reviewers and editor for taking the time to check out and provide feedback on our SedEdu educational software platform. The reviewers raised a handful of issues and suggestions in their review checklists and specific issues in the repository. We have directly responded to the direct comments, and responded to comments in the checklist where appropriate.

@labarba  please let me know how to proceed from here. Thank you all again! :smile:

# reviewer comments

## hernanrr review

> Minor installation issue: https://github.com/sededu/sededu/issues/98

Thank you for pointing this out, I've addressed this issue directly there. See related [SwansonHysell review issue](https://github.com/sededu/sededu/issues/99).

> There do not seem to be any software usage examples. This notwithstanding, the software, with its few buttons and sliders, is intuitive to use. Some, but not all, of the modules have worksheets that encourage and guide the the use of relevant sliders. The file [using_sededu.md](https://github.com/sededu/sededu/blob/develop/docs/using_sededu.md) is still mostly unpopulated and, thus, does not provide usage examples.

Thank you for pointing this out, we have updated the referenced page with a jargon-free description of the software, though we agree that the software is very intuitive.

## Swanson-Hysell review

> Regarding installation, I have raised this issue regarding the conda installation:
https://github.com/sededu/sededu/issues/99
> I have been able to install with pip.

Thank you for pointing this out, I've addressed this issue directly there. See related [hernanrr review issue](https://github.com/sededu/sededu/issues/98).

> The conda install is out of date with the current Python version available via Anaconda. The pip install worked just fine. While I appreciate that conda has made installation relatively straight-forward, I don't know if it rises to the level of simplicity implied here:
> > Integrating educational infrastructure into the classroom is challenging, in part, due to logistical considerations for installing software; these concerns are alleviated with SedEdu’s simple installation procedure

> The statement of need is very general and can be improved. It cites the benefit of active learning in general without highlighting the level of student that is being targeted or in what context the software is best applied (e.g. in class, in lab, for problem sets).

Thank you for this comment. This paper is about the SedEdu platform, and not about the modules. We therefore wrote the Statement of Need section to emphasize the niche for the software delivery mechanism rather than how the modules can be used. We disagree that this Statement of Need is general in that regard, because we specifically describe how SedEdu benefits educators and students (active learning) and researchers (teaching, broader impacts, existing coding skills). Nevertheless, we added/revised some text in the Statement of Need to also explain the need for the modules themselves.

> Given that the software does not expose the students to any code and student interaction is fully through a GUI interface, it is unclear why the developers haven't compiled the software so that it can be downloaded and run as an executable.

> This installation issue points to an overall issue related to the package. I can understand the pedagogical goal of giving students toy models where they can play around with variables and view outputs which is the design of the modules. However, if this is the sole goal, why not go ahead and do the work that makes the program be an executable binary. By doing so install issues would largely be circumvented.

Thank you for this suggestion. A goal during developing SedEdu was cross-platform support. It is, at a minimum, _very difficult_ to have the software work across three major operating systems. Providing an executable binary for each platform and system architecture is well beyond my experience. Providing a pip wheel and conda builds is something I've managed to cobble together, and achieves the goal of cross-platform support.

I am aware of some tools to build windows binaries, and given my anecdotal observation that Windows is the mode of platforms in undergraduate computer labs, this could be a viable option to support this use-case. Another option would be to convert the modules to run on a web server and be accessed via a browser without any installation required; but, I'm not sure if these tools fully exist yet for the interactivity in the SedEdu modules. Regardless, these would be future endeavors, so I have not taken any action to package the SedEdu software differently.

> In an instructional setting, if I am having students do the work of getting a full Python install with the necessary packages, I then want the students to get some exposure to code and the practices of scientific computing. However, in this package that is all hidden behind the GUI. It seems like it would be nice to have a version where the functionality is exposed in a way that students see some code (i.e. in Jupyter notebooks).

Thank you for this suggestion. Incorporating code in classrooms is awesome, and something I do in my teaching as well. But, these modules are complicated, and well above the level of undergraduate teaching that the modules target, so exposing the code is not realistic. I do like the idea of taking components of these modules and simplifying to create classroom coding activities. Lastly, this paper is technically on the SedEdu platform, rather than the modules. So, I have not taken any action to expose the module codes.
