### 2022-03-22 #129 @Swanson-Hysell
Hi @labarba, where can we find the updated manuscript? It makes sense for me to have a look at the text that they describe having revised.
