### 2022-03-22 #129 @amoodie
@Swanson-Hysell fwiw, the changes were pretty minor. I should have included a link to the diff, sorry about that.
https://github.com/sededu/sededu/commit/09bb90b57784da2b8d26a0abc9cd996dde17bf74
