### 2022-03-23 #129 @Swanson-Hysell
One of the main response to my input as that the paper is on the SedEdu platform rather than the underlying modules. Overall, this distinction to be content agnostic leaves little to comment on as the focus is then on a GUI dialog that opens up interactive GUIs where the user can vary parameters.

A positive comment that I made that was not addressed is:
> The example module is probably the best documentation of the software structure. https://github.com/sededu/example-module_sededu It could be valuable to link to it in the paper.
Could bear mentioning this documentation?

As I expressed in my original review, the project seems well-situated to be compiled as an executable since students are not expected to engage with code, I understand that doing so is non-trivial and it is fine that the authors say it is outside the scope. Nevertheless, the majority of computer users would not consider the steps necessary to run this software (i.e. full Python install and package installation) as 'simple' or alleviating logistical challenges; rather they consider downloading and double-clicking an executable to be 'simple.' That is ok, there are reasons to take on such logistical challenges to curate educational experiences, but the simplicity remains overstated in the manuscript in this passage:
> Integrating educational infrastructure into the classroom is challenging, in part, due to logistical considerations for installing software; these concerns are alleviated with SedEdu’s simple installation procedure,

Good to hear the conda install issue has been addressed.
