### 2022-03-23 fdc2d51
clarify installation approach, and add example as reference.
