+++ .jose/paper.bib
@@ -93,3 +93,11 @@
 	year = {2008},
 	pages = {3717--3725}
 }
+
+@misc{sededu_example_nodate,
+	title = {Example Module for SedEdu},
+	howpublished = {online},
+	author = {SedEdu},
+	url = {https://github.com/sededu/example-module_sededu},
+	year={2022}
+}
