### 2022-03-23 #129 @amoodie
Thank you for that positive comment. I apologize for not addressing it directly; it slipped past me. I did consider adding it in the last revision, but was not sure how to add it since it is not a peer reviewed reference. Anyway, I agree that it is useful, so I added it as a reference. A sentence in the Summary section now reads:
> Robust developer instructions included in the repository describe the workflow to develop and integrate a module into SedEdu [@sededu_example_nodate].

Your point on installation is taken. I have revised this text to clarify that SedEdu packages both the modules and infrastructure together, so that there is a single installation point. This sentence is revised as:
> Integrating educational infrastructure into the classroom is challenging, in part, due to logistical demand of installing multiple activities and modules; SedEdu provides a single installation that includes modules themselves and infrastructure to manage learning modules and activities. SedEdu is installed via Python package managers `pip` and `conda` (described in the project repository documentation), and follows established conventions that make installation straightforward for Python users.

Here is the diff: https://github.com/sededu/sededu/commit/fdc2d5133f4f196ad420b2280e5f4a9e56682e77
