### 2022-03-30 #129 @labarba
Thanks, @Swanson-Hysell — would you mind checking off your pending items on the checklist, for tidy closing of this review?
