### 2022-08-11 #129 @labarba
Oops. The final steps before recommending publication are:
- issue a tagged release on the target repository and tell us the version number here
- make an archive on Zenodo and report the DOI here
