### 2022-09-02 #129 @amoodie
Hi @labarba, I've tagged a release v1.1.10a and archived it to Zenodo with doi 10.5281/zenodo.7044896. Thanks!
