### 2022-09-12 #129 @labarba
@amoodie We like to ask authors to edit the metadata of the Zenodo deposit so the title and author list match the JOSS paper. Can you do that?
