### 2022-10-11 #129 @amoodie
Hi @labarba,

Great, and thank you for seeing this paper through to the end!

Unfortunately, [it looks like the pdf may not have deposited correctly](https://jose.theoj.org/papers/10.21105/jose.00129).

Please let me know if I should ping someone else to get that sorted. Thanks!
