### 2022-10-11 #129 @labarba
@openjournals/dev — could you take a look at this recently published JOSE paper?

I did get an email from GitHub that said:

> The page build failed for the `master` branch with the following error:
>
> Unable to sync pages directory. Please try again later.
>
> For information on troubleshooting Jekyll see:
>
> https://docs.github.com/articles/troubleshooting-jekyll-builds
