### 2022-10-12 #129 @arfon
> Ah shoot, that might be my fault. We changed the main branch name from master to develop at some point. It [looks like the pandoc build worked](https://github.com/openjournals/jose-papers/blob/93f767b4ff9f9601f3d0f349cfa967e613146394/jose.00129/10.21105.jose.00129.pdf).

@amoodie – it's not your fault. It's a failure on the `openjournals/jose-papers` repo. @labarba – I've noticed a few issues with GitHub Pages this last week – I just triggered a rebuild of the repo to see if that fixes it.
